User profile logic, renamed user service to user account service, add…

+++ backend/src/main/kotlin/org/aleks616/shrendar/band/controller/BandController.kt
@@ -97,7 +97,7 @@ class BandController (
 
     //WIKI ARTIST PAGE 2/2
     @GetMapping("/artist/{id}")
-    fun getBandsByArtistId(@PathVariable id:Int):List<ArtistBandsHistoryDto>{
+    fun getBandsByArtistId(@PathVariable id:Long):List<ArtistBandsHistoryDto>{
         return bandsMemberService.getBandsByArtistId(id)
     }
